Support TV Typescript types properly in 0.71
