+++ types/index.d.ts
@@ -157,6 +157,8 @@ export * from './public/Insets';
 export * from './public/ReactNativeRenderer';
 export * from './public/ReactNativeTypes';
 
+export * from './public/ReactNativeTVTypes';
+
 import type {ErrorUtils} from '../Libraries/vendor/core/ErrorUtils';
 
 declare global {
